Uso de condicional que permite verificar si un archivo existe para borrarlo

+++ files/delete.dart
@@ -4,7 +4,15 @@ void main() {
   /// Abrir archivo
   File ArchivoABorrar = new File("prueba.txt");
 
-  /// Borrar archivo
-  ArchivoABorrar.deleteSync();
-  print("Archivo borrado");
+  /// Comprobar si existe el archivo
+  if (ArchivoABorrar.existsSync()) {
+    print("Archivo encontrado");
+
+    /// Borrar archivo
+    ArchivoABorrar.deleteSync();
+    print("Archivo borrado");
+  } else {
+    print("Archivo no encontrado");
+    return;
+  }
 }
